chore: extract functions into own files
