+++ packages/shared/lib/core/wallet/utils/transactions/index.ts
@@ -1,6 +1,7 @@
-export * from './outputContainsValue'
+export * from './getDirectionFromTransaction'
 export * from './getFoundryOutputFromTransaction'
 export * from './getMainOutputFromTransaction'
 export * from './getNftOutputFromTransaction'
 export * from './getSenderFromTransaction'
 export * from './getSenderAddressFromInputs'
+export * from './outputContainsValue'
